Moving test to pcc dir and adjusting the conftest.py conflict; Fixing licence headers; Refactoring MoeGatePrefill
