--- models/demos/deepseek_v3_d_p/utils/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
-# SPDX-License-Identifier: Apache-2.0
